Clean up boilerplate SE Eventlist code.
This commit tracked me refactoring Streamelements User Eventlist code
that was just... so repetitive. This commit worked to DRYify some of
that code so both the "onWidgetLoad" and "onEventReceived" events are
using the same logic to populate the event list HTML rather than each of
them having their own long if/else statements.